* release/0.9.0: (29 commits)
  Fix running test_trans_serial on Cray with MPI=ON
  Require ecbuild 2.5, eckit 0.15, fckit 0.3, transi 0.4
  Version 0.9.0
  Fix warnings in documentation code
  add real32 for rank4 halo exchange
  ATLAS-73 Rename unit-test target
  ATLAS-73 Fortran interface
  ATLAS-73 Fixes and unit-test
  ATLAS-73 Expose trans function to compute derivatives from spectral (untested)
  ATLAS-77 Spectral norms, Atlas interface
  Flush Log buffers when atlas_finalize() is called
  Fix sign comparison warnings
  Remove pragma's
  Fix sign comparison warnings
  Update copyright headers for 2017
  ECBUILD-293 Add description for ecKit
  ECBUILD-293 Remove atlas prefix from feature descriptions
  ECKIT-197: prefer uint_ from C99 standard
  ECKIT-191: interface change for float and double compare
  Simplification of PointKdTree
  ...

+++ doc/user-guide/core-functionalities/functionspace/StructuredColumns.cc
@@ -31,7 +31,7 @@ int main(int argc, char *argv[])
     Grid::Ptr grid (Grid::create( "N32" ));
 
     // Number of points in the grid
-    int const nb_nodes = grid->npts();
+    size_t const nb_nodes = grid->npts();
 
     // Generate functionspace associated to grid
     StructuredColumns::Ptr
@@ -52,9 +52,9 @@ int main(int argc, char *argv[])
     Field::Ptr field_scalar1(fs_rgp->createField<double>("scalar1"));
     ArrayView <double,1> scalar1(*field_scalar1);
 
-    for (int jlat = 0; jlat < fs_rgp->nlat(); ++jlat)
+    for (size_t jlat = 0; jlat < fs_rgp->nlat(); ++jlat)
     {
-        for (int jlon = 0; jlon < fs_rgp->nlon(jlat); ++jlon)
+        for (size_t jlon = 0; jlon < fs_rgp->nlon(jlat); ++jlon)
         {
             double zlat = fs_rgp->lat(jlat);
             zlat = zlat * deg2rad;
@@ -78,7 +78,7 @@ int main(int argc, char *argv[])
       // Generate visualisation mesh associated to grid
       atlas::mesh::generators::Structured meshgenerator;
       Mesh::Ptr mesh (meshgenerator.generate(*grid));
-      
+
       Gmsh gmsh("scalar1.msh");
       gmsh.write(*mesh);
       gmsh.write(*field_scalar1);
