+++ src/atlas/mesh/actions/BuildHalo.cc
@@ -1,5 +1,5 @@
 /*
- * (C) Copyright 1996-2016 ECMWF.
+ * (C) Copyright 1996-2017 ECMWF.
  *
  * This software is licensed under the terms of the Apache Licence Version 2.0
  * which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
@@ -526,7 +526,7 @@ public:
       buf.elem_part   [p][jelem] = elem_part(ielem);
       buf.elem_type   [p][jelem] = mesh.cells().type_idx(ielem);
       std::vector<double> crds(elem_nodes->cols(ielem)*2);
-      for( int jnode=0; jnode<elem_nodes->cols(ielem); ++jnode)
+      for(size_t jnode=0; jnode<elem_nodes->cols(ielem); ++jnode)
       {
         double crd[] = { lonlat( (*elem_nodes)(ielem,jnode),internals::LON) , lonlat( (*elem_nodes)(ielem,jnode),internals::LAT) };
         transform(crd,-1);
